requirements.txt
scikit-learn
pandas
numpy
biopython
pytest
flask
jupyterlab
matplotlib
celery==5.1.1
click==7.1.2
redis
python-dotenv
-e .